Enhance the Ppr class
Now there is only Ppr and PprDebug.
I expect PprDebug to evolve more options in the future.

I removed the choice of different output types.
To output HTML (resp. XML), it is more appropriate
to use a HTML (resp. XML) combinator library.

Added a number of instances for built-in types.
I removed the [] and Set instances. These are
too specific to belong in the framework, as there
 is no single reasonable choice (e.g. you can print
a list elements in [1,2,3,..]  format or vertically
using a line for each element) and which one to use
should be a decision of the client.