Extended the Processor class with an applySearch method
Some processors return a choice of problems.
So far we have modeled this implicitly in the Proof Tree
with the Or nodes.
The applySearch method provides a explicit way to
model it.

In the long term we should remove Or and MPlus nodes from
the proof tree and approach search separately