redefine the Processor class
The former definition was too rigid to be useful.
It didn't allow a processor to change the type of
a trs. This can be necessary when transforming a
problem type to another.
For instance, when transforming an OSDPRewriting
problem to a DPRewriting problem, the type of the
ids in the TRS needs to be changed.

The new definition is equivalent to the original
definition in the 'master' branch of muterm-framework