Revert "Add a missing Pretty instance for Free"
This reverts commit 04b7d0022f7252cbc39991e507b1b10656c1fade.

Turns out that declaring the instance produces a conflict
in Narradar, since the same instance is declared in a
dependency of Narradar.

The solution adopted is to remove the instance and let
clients of muterm-framework declare the instance themselves.
Another option would be moving the instance to
control-monad-free, but that involves adding a
rather unjustified dependency from control-monad-free
on prettyclass,