pprProofFailures
This is a refactoring of a previous patch including support for
printing the contents of a search node.
Since a search node can contain multiple potential branches,
just printing them all is not what we want.
We want to evaluate the proof and print only the succesful branch.
However this only holds for a successful proof,
whereas the main use csae of displaying Search nodes is for reporting
failed proofs.
In this case what we want is to display only the part of the
tree that has been evaluated before deciding it has no solution.
This is exactly what sliceWorkDone in the GraphViz module is for.

This patch renames sliceWorkDone to sliceProof moving it to the
Proof module. In addition, the PprProofSearch class is dismissed
and replaced by a function pprProofFailures. pprProofFailures slices
the proof and then renders it. It requires that the MonadPlus used
in the proof is Foldable.